test(dgw): move injection helpers into testsuite lib
Move GatewayProc, provisioning, fake RDP target, mock KDC/RDP, CredSSP
client leg, agent harness, and TLS fixtures from the test files into
testsuite/src so the abstractions are self-contained, per review.

Add empty INTENT.md placeholders to be filled by hand.

Issue: DVLS-14697
