+++ testsuite/src/lib.rs
@@ -8,3 +8,5 @@ pub mod cli;
 pub mod dgw_config;
 pub mod mcp_client;
 pub mod mcp_server;
+pub mod rdp_injection;
+pub mod tls_fixtures;
