+++ testsuite/tests/cli/dgw/cred_injection_kdc.rs
@@ -7,965 +7,31 @@
 //! RDCleanPath coverage drives the public `ironrdp-agent` 0.1.0 CLI (`cargo install
 //! ironrdp-agent --version 0.1.0`) over `ws://127.0.0.1/jet/rdp`.
 
-use std::path::{Path, PathBuf};
-use std::process::Stdio;
-use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-use std::sync::{Arc, Mutex};
+use std::sync::Mutex;
 use std::time::{Duration, Instant};
 
 use anyhow::Context as _;
-use ironrdp_connector::sspi;
-use ironrdp_connector::sspi::generator::GeneratorState;
-use ironrdp_pdu::nego::{
-    ConnectionConfirm, ConnectionRequest, NegoRequestData, RequestFlags, ResponseFlags, SecurityProtocol,
+use testsuite::rdp_injection::agent::{
+    agent_query_logs, connect_ironrdp_rdcleanpath, ironrdp_agent_endpoint, require_ironrdp_agent, start_ironrdp_daemon,
 };
-use ironrdp_pdu::x224::X224;
-use ironrdp_tokio::{FramedWrite as _, TokioFramed};
-use picky_krb::data_types::PrincipalName;
-use picky_krb::messages::{AsRep, AsReq, KdcProxyMessage, KrbError, TgsRep, TgsReq};
-use rustls::pki_types::pem::PemObject as _;
-use rustls::pki_types::{CertificateDer, PrivateKeyDer, ServerName};
-use rustls::{ClientConfig, ServerConfig};
-use tokio::io::{AsyncBufReadExt as _, AsyncReadExt as _, AsyncWriteExt as _, BufReader};
-use tokio::net::{TcpListener, TcpStream};
-use x509_cert::der::Decode as _;
-
-use super::cred_injection::{
-    FORWARD_LOG, GatewayProc, INJECT_LOG, KERBEROS_TARGET_USER, MISSING_LOG, PROXY_KERBEROS_USER, PROXY_PASSWORD,
-    PROXY_USER, TARGET_PASSWORD, TARGET_USER, encode_pcb, next_id, provision_credentials, provision_mapping,
-    unsigned_jws,
+use testsuite::rdp_injection::credssp::{
+    complete_client_credssp, complete_ntlm_credssp, complete_raw_ntlm_credssp, connect_ntlm_client,
 };
-use super::tls_fixtures::{CERT_PEM, KEY_PEM};
-
-const REALM: &str = "EXAMPLE.INVALID";
-// sspi-rs downgrades Negotiate to NTLM when the SPN host is an IP address.
-const SERVICE_HOST: &str = "localhost";
-const KRBTGT_KEY: [u8; 32] = [0x11; 32];
-const TERMSRV_KEY: [u8; 32] = [0x22; 32];
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-enum ObservedKdcReq {
-    As { cname: String, realm: String },
-    Tgs { sname: Vec<String>, realm: String },
-    Other,
-}
-
-struct MockKdc {
-    port: u16,
-    exchanges: Arc<AtomicUsize>,
-    requests: Arc<Mutex<Vec<ObservedKdcReq>>>,
-}
-
-impl MockKdc {
-    async fn start() -> anyhow::Result<Self> {
-        let listener = TcpListener::bind("127.0.0.1:0").await.context("bind mock KDC")?;
-        let port = listener.local_addr().context("mock KDC local_addr")?.port();
-        let exchanges = Arc::new(AtomicUsize::new(0));
-        let requests = Arc::new(Mutex::new(Vec::new()));
-        let exchanges_task = Arc::clone(&exchanges);
-        let requests_task = Arc::clone(&requests);
-        let config = kdc_config();
-
-        tokio::spawn(async move {
-            loop {
-                let Ok((stream, _)) = listener.accept().await else {
-                    break;
-                };
-                let config = config.clone();
-                let exchanges = Arc::clone(&exchanges_task);
-                let requests = Arc::clone(&requests_task);
-                tokio::spawn(async move {
-                    match serve_kdc_exchange(stream, &config, &requests).await {
-                        Ok(()) => {
-                            exchanges.fetch_add(1, Ordering::SeqCst);
-                        }
-                        Err(error) => eprintln!("mock KDC exchange failed: {error:#}"),
-                    }
-                });
-            }
-        });
-
-        Ok(Self {
-            port,
-            exchanges,
-            requests,
-        })
-    }
-
-    fn url(&self) -> String {
-        format!("tcp://127.0.0.1:{}", self.port)
-    }
-
-    fn exchanges(&self) -> usize {
-        self.exchanges.load(Ordering::SeqCst)
-    }
-
-    fn requests(&self) -> Vec<ObservedKdcReq> {
-        self.requests.lock().expect("kdc request mutex").clone()
-    }
-}
-
-fn kdc_config() -> kdc::config::KerberosServer {
-    let username = format!("administrator@{REALM}");
-    kdc::config::KerberosServer {
-        realm: REALM.to_owned(),
-        users: vec![kdc::config::DomainUser {
-            username,
-            password: TARGET_PASSWORD.to_owned(),
-            salt: format!("{}administrator", REALM.to_ascii_uppercase()),
-        }],
-        max_time_skew: 300,
-        krbtgt_key: KRBTGT_KEY.to_vec(),
-        ticket_decryption_key: Some(TERMSRV_KEY.to_vec()),
-        service_user: None,
-    }
-}
-
-async fn serve_kdc_exchange(
-    mut stream: TcpStream,
-    config: &kdc::config::KerberosServer,
-    requests: &Mutex<Vec<ObservedKdcReq>>,
-) -> anyhow::Result<()> {
-    let mut len_buf = [0u8; 4];
-    stream.read_exact(&mut len_buf).await.context("read KDC length")?;
-    let len = usize::try_from(u32::from_be_bytes(len_buf)).context("KDC length")?;
-    let mut body = vec![0u8; len];
-    stream.read_exact(&mut body).await.context("read KDC body")?;
-
-    requests.lock().expect("kdc request mutex").push(observe_kdc_req(&body));
-
-    let mut raw = Vec::with_capacity(4 + len);
-    raw.extend_from_slice(&len_buf);
-    raw.extend_from_slice(&body);
-
-    let request = KdcProxyMessage::from_raw_kerb_message(&raw).context("wrap KDC TCP payload")?;
-    let reply = kdc::handle_kdc_proxy_message(request, config, SERVICE_HOST).context("handle KDC message")?;
-    stream
-        .write_all(&reply.kerb_message.0.0)
-        .await
-        .context("write KDC reply")?;
-    Ok(())
-}
-
-fn principal_strings(name: &PrincipalName) -> Vec<String> {
-    name.name_string.0.0.iter().map(|part| part.0.to_string()).collect()
-}
-
-fn observe_kdc_req(body: &[u8]) -> ObservedKdcReq {
-    if let Ok(as_req) = picky_asn1_der::from_bytes::<AsReq>(body) {
-        let req = &as_req.0.req_body.0;
-        let cname = req
-            .cname
-            .0
-            .as_ref()
-            .map(|name| principal_strings(&name.0).join("/"))
-            .unwrap_or_default();
-        return ObservedKdcReq::As {
-            cname,
-            realm: req.realm.0.to_string(),
-        };
-    }
-    if let Ok(tgs_req) = picky_asn1_der::from_bytes::<TgsReq>(body) {
-        let req = &tgs_req.0.req_body.0;
-        let sname = req
-            .sname
-            .0
-            .as_ref()
-            .map(|name| principal_strings(&name.0))
-            .unwrap_or_default();
-        return ObservedKdcReq::Tgs {
-            sname,
-            realm: req.realm.0.to_string(),
-        };
-    }
-    ObservedKdcReq::Other
-}
-
-fn observe_kdc_reply(body: &[u8]) -> ObservedKdcReply {
-    let krb = body.get(4..).unwrap_or(body);
-    if picky_asn1_der::from_bytes::<AsRep>(krb).is_ok() {
-        ObservedKdcReply::AsRep
-    } else if picky_asn1_der::from_bytes::<TgsRep>(krb).is_ok() {
-        ObservedKdcReply::TgsRep
-    } else if picky_asn1_der::from_bytes::<KrbError>(krb).is_ok() {
-        ObservedKdcReply::KrbError
-    } else {
-        ObservedKdcReply::Other
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-enum ObservedKdcReply {
-    AsRep,
-    TgsRep,
-    KrbError,
-    Other,
-}
-
-#[derive(Clone)]
-enum MockRdpMode {
-    Kerberos { kdc_url: Option<String> },
-    Ntlm,
-}
-
-struct MockRdp {
-    port: u16,
-    credssp_ok: Arc<AtomicBool>,
-    finished_account: Arc<Mutex<Option<String>>>,
-    cookies: Arc<Mutex<Vec<String>>>,
-}
-
-impl MockRdp {
-    async fn start_kerberos(kdc_url: String) -> anyhow::Result<Self> {
-        Self::start(MockRdpMode::Kerberos { kdc_url: Some(kdc_url) }).await
-    }
-
-    async fn start_ntlm() -> anyhow::Result<Self> {
-        Self::start(MockRdpMode::Ntlm).await
-    }
-
-    async fn start(mode: MockRdpMode) -> anyhow::Result<Self> {
-        install_crypto_provider();
-        // Dual-stack so Windows `localhost` (IPv6 first) still hits the fake server.
-        let listener = match TcpListener::bind("[::]:0").await {
-            Ok(listener) => listener,
-            Err(_) => TcpListener::bind("127.0.0.1:0").await.context("bind mock RDP")?,
-        };
-        let port = listener.local_addr().context("mock RDP local_addr")?.port();
-        let credssp_ok = Arc::new(AtomicBool::new(false));
-        let finished_account = Arc::new(Mutex::new(None));
-        let cookies = Arc::new(Mutex::new(Vec::new()));
-        let credssp_ok_task = Arc::clone(&credssp_ok);
-        let finished_account_task = Arc::clone(&finished_account);
-        let cookies_task = Arc::clone(&cookies);
-        let acceptor = tls_acceptor()?;
-        let public_key = server_public_key()?;
-
-        tokio::spawn(async move {
-            loop {
-                let Ok((stream, peer)) = listener.accept().await else {
-                    break;
-                };
-                let acceptor = acceptor.clone();
-                let public_key = public_key.clone();
-                let credssp_ok = Arc::clone(&credssp_ok_task);
-                let finished_account = Arc::clone(&finished_account_task);
-                let cookies = Arc::clone(&cookies_task);
-                let mode = mode.clone();
-                tokio::spawn(async move {
-                    let result = match &mode {
-                        MockRdpMode::Kerberos { kdc_url } => {
-                            accept_kerberos_rdp(
-                                stream,
-                                peer,
-                                acceptor,
-                                public_key,
-                                kdc_url.as_deref(),
-                                &cookies,
-                                &finished_account,
-                            )
-                            .await
-                        }
-                        MockRdpMode::Ntlm => {
-                            accept_ntlm_rdp(stream, peer, acceptor, public_key, &cookies, &finished_account).await
-                        }
-                    };
-                    match result {
-                        Ok(()) => credssp_ok.store(true, Ordering::SeqCst),
-                        Err(error) => eprintln!("mock RDP CredSSP failed: {error:#}"),
-                    }
-                });
-            }
-        });
-
-        Ok(Self {
-            port,
-            credssp_ok,
-            finished_account,
-            cookies,
-        })
-    }
-
-    fn credssp_ok(&self) -> bool {
-        self.credssp_ok.load(Ordering::SeqCst)
-    }
-
-    fn finished_account(&self) -> Option<String> {
-        self.finished_account.lock().expect("finished account mutex").clone()
-    }
-
-    fn cookies(&self) -> Vec<String> {
-        self.cookies.lock().expect("cookie mutex").clone()
-    }
-
-    async fn wait_credssp(&self) -> anyhow::Result<()> {
-        let deadline = Instant::now() + Duration::from_secs(30);
-        loop {
-            if self.credssp_ok() {
-                return Ok(());
-            }
-            if Instant::now() >= deadline {
-                anyhow::bail!("timed out waiting for Kerberos CredSSP on mock RDP");
-            }
-            tokio::time::sleep(Duration::from_millis(50)).await;
-        }
-    }
-}
-
-async fn accept_kerberos_rdp(
-    stream: TcpStream,
-    peer: std::net::SocketAddr,
-    acceptor: tokio_rustls::TlsAcceptor,
-    public_key: Vec<u8>,
-    kdc_url: Option<&str>,
-    cookies: &Mutex<Vec<String>>,
-    finished_account: &Mutex<Option<String>>,
-) -> anyhow::Result<()> {
-    let mut framed = TokioFramed::new(stream);
-    let (_, request) = framed.read_pdu().await.context("read X.224 CR")?;
-    let cr: X224<ConnectionRequest> = ironrdp_core::decode(&request).context("decode X.224 CR")?;
-    record_cookie(&cr, cookies);
-
-    let confirm = X224(ConnectionConfirm::Response {
-        flags: ResponseFlags::empty(),
-        protocol: SecurityProtocol::HYBRID,
-    });
-    framed
-        .write_all(&ironrdp_core::encode_vec(&confirm).context("encode X.224 CC")?)
-        .await
-        .context("write X.224 CC")?;
-
-    let tcp = framed.into_inner_no_leftover();
-    let tls = acceptor.accept(tcp).await.context("TLS accept")?;
-    let mut framed = TokioFramed::new(tls);
-
-    let identity = sspi::AuthIdentity {
-        username: sspi::Username::parse(KERBEROS_TARGET_USER).context("parse target username")?,
-        password: TARGET_PASSWORD.to_owned().into(),
-    };
-    let kerberos_config = sspi::KerberosServerConfig {
-        kerberos_config: sspi::KerberosConfig {
-            kdc_url: kdc_url
-                .map(|url| url.parse())
-                .transpose()
-                .context("parse mock KDC URL")?,
-            client_computer_name: peer.to_string(),
-        },
-        server_properties: sspi::kerberos::ServerProperties::new(
-            &["TERMSRV", SERVICE_HOST],
-            Some(sspi::CredentialsBuffers::AuthIdentity(
-                sspi::AuthIdentityBuffers::from_utf8(identity.username.account_name(), REALM, TARGET_PASSWORD),
-            )),
-            Duration::from_secs(300),
-            Some(sspi::Secret::new(TERMSRV_KEY.to_vec())),
-        )
-        .context("Kerberos server properties")?,
-    };
-
-    let mut server = sspi::credssp::CredSspServer::new(
-        public_key,
-        IdentityProxy(identity),
-        sspi::credssp::ServerMode::Negotiate(sspi::NegotiateConfig::new(
-            Box::new(kerberos_config),
-            Some("kerberos,!ntlm".to_owned()),
-            peer.to_string(),
-        )),
-    )
-    .context("init Kerberos-only CredSSP server")?;
-
-    let hint = TsRequestHint;
-    let mut buf = ironrdp_pdu::WriteBuf::new();
-    for _ in 0..6 {
-        let pdu = framed.read_by_hint(&hint).await.context("read CredSSP TSRequest")?;
-        let ts_request = sspi::credssp::TsRequest::from_buffer(&pdu).context("decode CredSSP")?;
-        let result = {
-            let mut generator = server.process(ts_request);
-            resolve_sspi_server(&mut generator)
-                .await
-                .map_err(|error| anyhow::anyhow!("mock RDP CredSSP: {error:?}"))?
-        };
-        match result {
-            sspi::credssp::ServerState::ReplyNeeded(outbound) => {
-                buf.clear();
-                let length = usize::from(outbound.buffer_len());
-                outbound
-                    .encode_ts_request(buf.unfilled_to(length))
-                    .context("encode server TSRequest")?;
-                buf.advance(length);
-                framed.write_all(&buf[..length]).await.context("write CredSSP")?;
-            }
-            sspi::credssp::ServerState::Finished(identity) => {
-                *finished_account.lock().expect("finished account mutex") =
-                    Some(identity.username.account_name().to_owned());
-                return Ok(());
-            }
-        }
-    }
-    anyhow::bail!("mock RDP CredSSP exceeded 6 round trips")
-}
-
-async fn accept_ntlm_rdp(
-    stream: TcpStream,
-    peer: std::net::SocketAddr,
-    acceptor: tokio_rustls::TlsAcceptor,
-    public_key: Vec<u8>,
-    cookies: &Mutex<Vec<String>>,
-    finished_account: &Mutex<Option<String>>,
-) -> anyhow::Result<()> {
-    let mut framed = TokioFramed::new(stream);
-    let (_, request) = framed.read_pdu().await.context("read X.224 CR")?;
-    let cr: X224<ConnectionRequest> = ironrdp_core::decode(&request).context("decode X.224 CR")?;
-    record_cookie(&cr, cookies);
-
-    let confirm = X224(ConnectionConfirm::Response {
-        flags: ResponseFlags::empty(),
-        protocol: SecurityProtocol::HYBRID,
-    });
-    framed
-        .write_all(&ironrdp_core::encode_vec(&confirm).context("encode X.224 CC")?)
-        .await
-        .context("write X.224 CC")?;
-
-    let tcp = framed.into_inner_no_leftover();
-    let tls = acceptor.accept(tcp).await.context("TLS accept")?;
-    let mut framed = TokioFramed::new(tls);
-
-    let identity = sspi::AuthIdentity {
-        username: sspi::Username::parse(TARGET_USER).context("parse NTLM target username")?,
-        password: TARGET_PASSWORD.to_owned().into(),
-    };
-    let mut server = sspi::credssp::CredSspServer::new(
-        public_key,
-        IdentityProxy(identity),
-        sspi::credssp::ServerMode::Ntlm(sspi::ntlm::NtlmConfig {
-            client_computer_name: Some(peer.to_string()),
-        }),
-    )
-    .context("init NTLM CredSSP server")?;
-
-    let hint = TsRequestHint;
-    let mut buf = ironrdp_pdu::WriteBuf::new();
-    for _ in 0..6 {
-        let pdu = framed.read_by_hint(&hint).await.context("read CredSSP TSRequest")?;
-        let ts_request = sspi::credssp::TsRequest::from_buffer(&pdu).context("decode CredSSP")?;
-        let result = {
-            let mut generator = server.process(ts_request);
-            resolve_sspi_server(&mut generator)
-                .await
-                .map_err(|error| anyhow::anyhow!("mock RDP NTLM CredSSP: {error:?}"))?
-        };
-        match result {
-            sspi::credssp::ServerState::ReplyNeeded(outbound) => {
-                buf.clear();
-                let length = usize::from(outbound.buffer_len());
-                outbound
-                    .encode_ts_request(buf.unfilled_to(length))
-                    .context("encode server TSRequest")?;
-                buf.advance(length);
-                framed.write_all(&buf[..length]).await.context("write CredSSP")?;
-            }
-            sspi::credssp::ServerState::Finished(identity) => {
-                *finished_account.lock().expect("finished account mutex") =
-                    Some(identity.username.account_name().to_owned());
-                return Ok(());
-            }
-        }
-    }
-    anyhow::bail!("mock RDP NTLM CredSSP exceeded 6 round trips")
-}
-
-fn record_cookie(cr: &X224<ConnectionRequest>, cookies: &Mutex<Vec<String>>) {
-    if let Some(NegoRequestData::Cookie(cookie)) = &cr.0.nego_data {
-        cookies.lock().expect("cookie mutex").push(cookie.0.clone());
-    }
-}
-
-async fn resolve_sspi_server(
-    generator: &mut sspi::generator::Generator<
-        '_,
-        sspi::generator::NetworkRequest,
-        sspi::Result<Vec<u8>>,
-        Result<sspi::credssp::ServerState, sspi::credssp::ServerError>,
-    >,
-) -> Result<sspi::credssp::ServerState, sspi::credssp::ServerError> {
-    let mut state = generator.start();
-    loop {
-        match state {
-            GeneratorState::Suspended(request) => {
-                let reply = send_kdc_tcp(&request)
-                    .await
-                    .map_err(|error| sspi::credssp::ServerError {
-                        ts_request: None,
-                        error: sspi::Error::new(sspi::ErrorKind::NoAuthenticatingAuthority, error),
-                    })?;
-                state = generator.resume(Ok(reply));
-            }
-            GeneratorState::Completed(result) => break result,
-        }
-    }
-}
-
-async fn send_kdc_tcp(request: &sspi::generator::NetworkRequest) -> anyhow::Result<Vec<u8>> {
-    let host = request.url.host_str().context("KDC host")?;
-    let port = request.url.port().unwrap_or(88);
-    let mut stream = TcpStream::connect((host, port)).await.context("connect mock KDC")?;
-    stream.write_all(&request.data).await.context("write KDC request")?;
-    let mut len_buf = [0u8; 4];
-    stream.read_exact(&mut len_buf).await.context("read KDC length")?;
-    let len = usize::try_from(u32::from_be_bytes(len_buf)).context("KDC length")?;
-    let mut body = vec![0u8; len];
-    stream.read_exact(&mut body).await.context("read KDC body")?;
-    let mut reply = Vec::with_capacity(4 + len);
-    reply.extend_from_slice(&len_buf);
-    reply.extend_from_slice(&body);
-    Ok(reply)
-}
-
-struct IdentityProxy(sspi::AuthIdentity);
-
-impl sspi::credssp::CredentialsProxy for IdentityProxy {
-    type AuthenticationData = sspi::AuthIdentity;
-
-    fn auth_data_by_user(&mut self, username: &sspi::Username) -> std::io::Result<Self::AuthenticationData> {
-        if username.account_name() != self.0.username.account_name() {
-            return Err(std::io::Error::other("invalid username"));
-        }
-        let mut data = self.0.clone();
-        data.username = username.clone();
-        Ok(data)
-    }
-
-    fn auth_data(&mut self) -> Result<Vec<Self::AuthenticationData>, std::io::Error> {
-        Ok(vec![self.0.clone()])
-    }
-}
-
-const HANDSHAKE_TIMEOUT: Duration = Duration::from_secs(20);
-
-async fn connect_ntlm_client(
-    gateway_tcp: u16,
-    association_jwt: &str,
-) -> anyhow::Result<tokio_rustls::client::TlsStream<TcpStream>> {
-    tokio::time::timeout(
-        HANDSHAKE_TIMEOUT,
-        connect_ntlm_client_inner(gateway_tcp, association_jwt),
-    )
-    .await
-    .context("timed out connecting NTLM client to Gateway")?
-}
-
-async fn connect_ntlm_client_inner(
-    gateway_tcp: u16,
-    association_jwt: &str,
-) -> anyhow::Result<tokio_rustls::client::TlsStream<TcpStream>> {
-    let mut stream = TcpStream::connect(("127.0.0.1", gateway_tcp))
-        .await
-        .context("connect gateway TCP")?;
-    stream
-        .write_all(&encode_pcb(association_jwt)?)
-        .await
-        .context("write PCB")?;
-    stream.write_all(&encode_hybrid_cr()?).await.context("write X.224 CR")?;
-    stream.flush().await.context("flush CR")?;
-
-    let mut framed = TokioFramed::new(stream);
-    let (_, confirm) = framed.read_pdu().await.context("read X.224 CC")?;
-    let confirm: X224<ConnectionConfirm> = ironrdp_core::decode(&confirm).context("decode X.224 CC")?;
-    anyhow::ensure!(
-        matches!(confirm.0, ConnectionConfirm::Response { protocol, .. } if protocol.contains(SecurityProtocol::HYBRID)),
-        "gateway did not confirm CredSSP: {confirm:?}"
-    );
-
-    let tcp = framed.into_inner_no_leftover();
-    let connector = dangerous_tls_connector();
-    let server_name = ServerName::try_from("localhost").map_err(|error| anyhow::anyhow!("{error}"))?;
-    connector.connect(server_name, tcp).await.context("TLS to gateway")
-}
-
-async fn complete_ntlm_credssp(tls: tokio_rustls::client::TlsStream<TcpStream>) -> anyhow::Result<()> {
-    complete_client_credssp(tls, PROXY_USER, None, false, None, None).await
-}
-
-async fn complete_raw_ntlm_credssp(tls: tokio_rustls::client::TlsStream<TcpStream>) -> anyhow::Result<()> {
-    complete_client_credssp(tls, PROXY_USER, None, true, None, None).await
-}
-
-async fn complete_client_credssp(
-    tls: tokio_rustls::client::TlsStream<TcpStream>,
-    username: &str,
-    kdc_proxy_url: Option<&str>,
-    raw_ntlm: bool,
-    proxy_replies: Option<&Mutex<Vec<ObservedKdcReply>>>,
-    proxy_requests: Option<&Mutex<Vec<ObservedKdcReq>>>,
-) -> anyhow::Result<()> {
-    tokio::time::timeout(
-        HANDSHAKE_TIMEOUT,
-        complete_client_credssp_inner(tls, username, kdc_proxy_url, raw_ntlm, proxy_replies, proxy_requests),
-    )
-    .await
-    .context("timed out completing client CredSSP")?
-}
-
-async fn complete_client_credssp_inner(
-    tls: tokio_rustls::client::TlsStream<TcpStream>,
-    username: &str,
-    kdc_proxy_url: Option<&str>,
-    raw_ntlm: bool,
-    proxy_replies: Option<&Mutex<Vec<ObservedKdcReply>>>,
-    proxy_requests: Option<&Mutex<Vec<ObservedKdcReq>>>,
-) -> anyhow::Result<()> {
-    use sspi::credssp::{ClientMode, ClientState, CredSspClient, CredSspMode, TsRequest};
-    use sspi::ntlm::NtlmConfig;
-
-    let public_key = peer_public_key(&tls)?;
-    let mut framed = TokioFramed::new(tls);
-    let identity = sspi::AuthIdentity {
-        username: sspi::Username::parse(username).context("parse client username")?,
-        password: PROXY_PASSWORD.to_owned().into(),
-    };
-    let client_mode = if let Some(kdc_url) = kdc_proxy_url {
-        ClientMode::Negotiate(sspi::NegotiateConfig::new(
-            Box::new(sspi::KerberosConfig {
-                kdc_url: Some(kdc_url.parse().context("parse KDC proxy URL")?),
-                client_computer_name: "cred-injection-e2e".to_owned(),
-            }),
-            Some("kerberos,!ntlm".to_owned()),
-            "cred-injection-e2e".to_owned(),
-        ))
-    } else if raw_ntlm {
-        // Gateway NTLM injection uses ServerMode::Ntlm, which rejects SPNEGO.
-        ClientMode::Ntlm(NtlmConfig {
-            client_computer_name: Some("cred-injection-e2e".to_owned()),
-        })
-    } else {
-        ClientMode::Negotiate(sspi::NegotiateConfig::new(
-            Box::new(NtlmConfig {
-                client_computer_name: Some("cred-injection-e2e".to_owned()),
-            }),
-            Some("ntlm,!kerberos,!pku2u".to_owned()),
-            "cred-injection-e2e".to_owned(),
-        ))
-    };
-    let mut client = CredSspClient::new(
-        public_key,
-        identity.into(),
-        CredSspMode::WithCredentials,
-        client_mode,
-        format!("TERMSRV/{SERVICE_HOST}"),
-    )
-    .context("init CredSSP client")?;
-
-    let mut ts_request = TsRequest::default();
-    let mut buf = ironrdp_pdu::WriteBuf::new();
-    let hint = TsRequestHint;
-
-    for _ in 0..8 {
-        let client_state = {
-            let mut generator = client.process(std::mem::take(&mut ts_request));
-            resolve_sspi_client(&mut generator, proxy_replies, proxy_requests).await?
-        };
-        let (outbound, finished) = match client_state {
-            ClientState::ReplyNeeded(request) => (request, false),
-            ClientState::FinalMessage(request) => (request, true),
-        };
-        buf.clear();
-        let length = usize::from(outbound.buffer_len());
-        outbound
-            .encode_ts_request(buf.unfilled_to(length))
-            .context("encode client TSRequest")?;
-        buf.advance(length);
-        framed.write_all(&buf[..length]).await.context("write client CredSSP")?;
-        if finished {
-            return Ok(());
-        }
-        let pdu = framed.read_by_hint(&hint).await.context("read server CredSSP")?;
-        ts_request = TsRequest::from_buffer(&pdu).context("decode server TSRequest")?;
-    }
-
-    anyhow::bail!("CredSSP exceeded 8 round trips")
-}
-
-async fn resolve_sspi_client(
-    generator: &mut sspi::generator::Generator<
-        '_,
-        sspi::generator::NetworkRequest,
-        sspi::Result<Vec<u8>>,
-        sspi::Result<sspi::credssp::ClientState>,
-    >,
-    proxy_replies: Option<&Mutex<Vec<ObservedKdcReply>>>,
-    proxy_requests: Option<&Mutex<Vec<ObservedKdcReq>>>,
-) -> anyhow::Result<sspi::credssp::ClientState> {
-    let mut state = generator.start();
-    loop {
-        match state {
-            GeneratorState::Suspended(request) => {
-                let reply = match request.url.scheme() {
-                    "tcp" | "udp" => send_kdc_tcp(&request).await?,
-                    "http" | "https" => send_kdc_http(&request, proxy_replies, proxy_requests).await?,
-                    other => anyhow::bail!("unsupported KDC scheme {other}: {}", request.url),
-                };
-                state = generator.resume(Ok(reply));
-            }
-            GeneratorState::Completed(result) => {
-                break result.map_err(|error| anyhow::anyhow!("client CredSSP: {error}"));
-            }
-        }
-    }
-}
-
-async fn send_kdc_http(
-    request: &sspi::generator::NetworkRequest,
-    proxy_replies: Option<&Mutex<Vec<ObservedKdcReply>>>,
-    proxy_requests: Option<&Mutex<Vec<ObservedKdcReq>>>,
-) -> anyhow::Result<Vec<u8>> {
-    let host = request.url.host_str().context("KDC proxy host")?;
-    let port = request.url.port_or_known_default().unwrap_or(80);
-    let path = if request.url.query().is_some() {
-        format!("{}?{}", request.url.path(), request.url.query().unwrap_or_default())
-    } else {
-        request.url.path().to_owned()
-    };
-    let mut stream = TcpStream::connect((host, port)).await.context("connect KDC proxy")?;
-    let header = format!(
-        "POST {path} HTTP/1.1\r\n\
-         Host: {host}:{port}\r\n\
-         Content-Type: application/octet-stream\r\n\
-         Content-Length: {}\r\n\
-         Connection: close\r\n\
-         \r\n",
-        request.data.len()
-    );
-    stream
-        .write_all(header.as_bytes())
-        .await
-        .context("write KDC proxy headers")?;
-    stream.write_all(&request.data).await.context("write KDC proxy body")?;
-    stream.flush().await.context("flush KDC proxy")?;
-    if let Ok(message) = KdcProxyMessage::from_raw(&request.data)
-        && let Some(log) = proxy_requests
-    {
-        let kerb = message.kerb_message.0.0.get(4..).unwrap_or(&message.kerb_message.0.0);
-        log.lock().expect("proxy request mutex").push(observe_kdc_req(kerb));
-    }
-
-    let mut reader = BufReader::new(stream);
-    let mut status_line = String::new();
-    reader
-        .read_line(&mut status_line)
-        .await
-        .context("read KDC proxy status")?;
-    anyhow::ensure!(
-        status_line.starts_with("HTTP/1.1 200") || status_line.starts_with("HTTP/1.0 200"),
-        "KDC proxy HTTP status was {status_line:?}"
-    );
-
-    let mut content_length = None;
-    loop {
-        let mut line = String::new();
-        reader.read_line(&mut line).await.context("read KDC proxy header")?;
-        if line == "\r\n" || line.is_empty() {
-            break;
-        }
-        if let Some(value) = line
-            .split_once(':')
-            .filter(|(name, _)| name.eq_ignore_ascii_case("content-length"))
-            .map(|(_, value)| value.trim().to_owned())
-        {
-            content_length = Some(value.parse::<usize>().context("parse KDC proxy Content-Length")?);
-        }
-    }
-
-    let buf = if let Some(len) = content_length {
-        let mut buf = vec![0u8; len];
-        tokio::io::AsyncReadExt::read_exact(&mut reader, &mut buf)
-            .await
-            .context("read KDC proxy body")?;
-        buf
-    } else {
-        let mut buf = Vec::new();
-        tokio::io::AsyncReadExt::read_to_end(&mut reader, &mut buf)
-            .await
-            .context("read KDC proxy eof body")?;
-        buf
-    };
-    if let Ok(message) = KdcProxyMessage::from_raw(&buf)
-        && let Some(log) = proxy_replies
-    {
-        log.lock()
-            .expect("proxy reply mutex")
-            .push(observe_kdc_reply(&message.kerb_message.0.0));
-    }
-    Ok(buf)
-}
-
-fn kdc_inject_token(association_jti: &str) -> anyhow::Result<String> {
-    unsigned_jws(
-        serde_json::json!({"alg":"RS256","typ":"JWT","cty":"KDC"}),
-        serde_json::json!({
-            "exp": 9_999_999_999i64,
-            "jet_cred_id": association_jti,
-            "jti": next_id(),
-        }),
-    )
-}
-
-fn kdc_proxy_url(http_port: u16, association_jti: &str) -> anyhow::Result<String> {
-    let token = kdc_inject_token(association_jti)?;
-    Ok(format!("http://127.0.0.1:{http_port}/jet/KdcProxy/{token}"))
-}
-
-#[derive(Debug)]
-struct TsRequestHint;
-
-impl ironrdp_pdu::PduHint for TsRequestHint {
-    fn find_size(&self, bytes: &[u8]) -> ironrdp_core::DecodeResult<Option<(bool, usize)>> {
-        match sspi::credssp::TsRequest::read_length(bytes) {
-            Ok(length) => Ok(Some((true, length))),
-            Err(error) if error.kind() == std::io::ErrorKind::UnexpectedEof => Ok(None),
-            Err(error) => Err(ironrdp_core::other_err!("TsRequestHint", source: error)),
-        }
-    }
-}
-
-fn association_token_for_host(jti: &str, jet_aid: &str, dest_port: u16, jet_reuse: u32) -> anyhow::Result<String> {
-    unsigned_jws(
-        serde_json::json!({"alg":"RS256","typ":"JWT","cty":"ASSOCIATION"}),
-        serde_json::json!({
-            "dst_hst": format!("{SERVICE_HOST}:{dest_port}"),
-            "exp": 9_999_999_999i64,
-            "jet_aid": jet_aid,
-            "jet_ap": "rdp",
-            "jet_cm": "fwd",
-            "jet_rec": "none",
-            "jet_reuse": jet_reuse,
-            "jti": jti,
-            "nbf": 0,
-        }),
-    )
-}
-
-fn encode_hybrid_cr() -> anyhow::Result<Vec<u8>> {
-    let pdu = X224(ConnectionRequest {
-        nego_data: Some(NegoRequestData::cookie(super::cred_injection::CLIENT_COOKIE.to_owned())),
-        flags: RequestFlags::empty(),
-        protocol: SecurityProtocol::HYBRID | SecurityProtocol::SSL,
-    });
-    ironrdp_core::encode_vec(&pdu).context("encode hybrid CR")
-}
-
-fn peer_public_key(tls: &tokio_rustls::client::TlsStream<TcpStream>) -> anyhow::Result<Vec<u8>> {
-    let cert = tls
-        .get_ref()
-        .1
-        .peer_certificates()
-        .and_then(|certs| certs.first())
-        .context("gateway TLS certificate missing")?;
-    extract_public_key(cert)
-}
-
-fn server_public_key() -> anyhow::Result<Vec<u8>> {
-    let cert = CertificateDer::from_pem_slice(CERT_PEM.as_bytes()).context("parse mock RDP cert")?;
-    extract_public_key(&cert)
-}
-
-fn extract_public_key(cert: &CertificateDer<'_>) -> anyhow::Result<Vec<u8>> {
-    let cert = x509_cert::Certificate::from_der(cert.as_ref()).context("parse X509")?;
-    let public_key = cert
-        .tbs_certificate()
-        .subject_public_key_info()
-        .subject_public_key
-        .as_bytes()
-        .context("unaligned subject public key")?
-        .to_owned();
-    Ok(public_key)
-}
-
-fn tls_acceptor() -> anyhow::Result<tokio_rustls::TlsAcceptor> {
-    let cert = CertificateDer::from_pem_slice(CERT_PEM.as_bytes()).context("parse cert PEM")?;
-    let key = PrivateKeyDer::from_pem_slice(KEY_PEM.as_bytes()).context("parse key PEM")?;
-    let config = ServerConfig::builder()
-        .with_no_client_auth()
-        .with_single_cert(vec![cert], key)
-        .context("TLS server config")?;
-    Ok(tokio_rustls::TlsAcceptor::from(Arc::new(config)))
-}
-
-fn dangerous_tls_connector() -> tokio_rustls::TlsConnector {
-    let mut config = ClientConfig::builder()
-        .dangerous()
-        .with_custom_certificate_verifier(Arc::new(NoCertificateVerification))
-        .with_no_client_auth();
-    config.resumption = rustls::client::Resumption::disabled();
-    tokio_rustls::TlsConnector::from(Arc::new(config))
-}
-
-fn install_crypto_provider() {
-    let _ = rustls::crypto::ring::default_provider().install_default();
-}
-
-#[derive(Debug)]
-struct NoCertificateVerification;
-
-impl rustls::client::danger::ServerCertVerifier for NoCertificateVerification {
-    fn verify_server_cert(
-        &self,
-        _: &CertificateDer<'_>,
-        _: &[CertificateDer<'_>],
-        _: &ServerName<'_>,
-        _: &[u8],
-        _: rustls::pki_types::UnixTime,
-    ) -> Result<rustls::client::danger::ServerCertVerified, rustls::Error> {
-        Ok(rustls::client::danger::ServerCertVerified::assertion())
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        _: &[u8],
-        _: &CertificateDer<'_>,
-        _: &rustls::DigitallySignedStruct,
-    ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
-        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        _: &[u8],
-        _: &CertificateDer<'_>,
-        _: &rustls::DigitallySignedStruct,
-    ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
-        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<rustls::SignatureScheme> {
-        vec![
-            rustls::SignatureScheme::RSA_PKCS1_SHA256,
-            rustls::SignatureScheme::ECDSA_NISTP256_SHA256,
-            rustls::SignatureScheme::RSA_PSS_SHA256,
-            rustls::SignatureScheme::ED25519,
-        ]
-    }
-}
-
-fn assert_target_kdc_as_and_tgs(kdc: &MockKdc) -> anyhow::Result<()> {
-    let reqs = kdc.requests();
-    anyhow::ensure!(
-        reqs.iter().any(|req| matches!(
-            req,
-            ObservedKdcReq::As { cname, realm }
-                if cname.eq_ignore_ascii_case("administrator") && realm.eq_ignore_ascii_case(REALM)
-        )),
-        "KDC must see AS-REQ cname=administrator realm={REALM}; requests={reqs:?}"
-    );
-    anyhow::ensure!(
-        reqs.iter().any(|req| matches!(
-            req,
-            ObservedKdcReq::Tgs { sname, realm }
-                if *sname == ["TERMSRV", SERVICE_HOST] && realm.eq_ignore_ascii_case(REALM)
-        )),
-        "KDC must see TGS-REQ sname=TERMSRV/{SERVICE_HOST} realm={REALM}; requests={reqs:?}"
-    );
-    Ok(())
-}
+use testsuite::rdp_injection::gateway::GatewayProc;
+use testsuite::rdp_injection::mock_kdc::{
+    MockKdc, ObservedKdcReply, ObservedKdcReq, RefusingKdc, assert_target_kdc_as_and_tgs,
+};
+use testsuite::rdp_injection::mock_rdp::MockRdp;
+use testsuite::rdp_injection::preflight::{provision_credentials, provision_mapping};
+use testsuite::rdp_injection::rdp::{FakeClosedTarget, encode_hybrid_cr, encode_pcb};
+use testsuite::rdp_injection::tls::install_crypto_provider;
+use testsuite::rdp_injection::tokens::{association_token_for_host, kdc_proxy_url, next_id};
+use testsuite::rdp_injection::{
+    FORWARD_LOG, INJECT_LOG, KERBEROS_TARGET_USER, MISSING_LOG, PROXY_KERBEROS_USER, PROXY_USER,
+    RDCLEANPATH_FORWARD_LOG, RDCLEANPATH_INJECT_LOG, REALM, SERVICE_HOST, TARGET_PASSWORD, TARGET_USER,
+};
+use tokio::io::AsyncWriteExt as _;
+use tokio::net::TcpStream;
 
 #[tokio::test]
 async fn kerberos_injection_completes_credssp_against_mock_kdc() -> anyhow::Result<()> {
@@ -1182,37 +248,6 @@ async fn kerberos_wrong_target_password_fails_closed() -> anyhow::Result<()> {
     Ok(())
 }
 
-struct RefusingKdc {
-    port: u16,
-    accepted: Arc<AtomicUsize>,
-}
-
-impl RefusingKdc {
-    async fn start() -> anyhow::Result<Self> {
-        let listener = TcpListener::bind("127.0.0.1:0").await.context("bind refusing KDC")?;
-        let port = listener.local_addr()?.port();
-        let accepted = Arc::new(AtomicUsize::new(0));
-        let accepted_task = Arc::clone(&accepted);
-        tokio::spawn(async move {
-            loop {
-                let Ok((_stream, _)) = listener.accept().await else {
-                    break;
-                };
-                accepted_task.fetch_add(1, Ordering::SeqCst);
-            }
-        });
-        Ok(Self { port, accepted })
-    }
-
-    fn url(&self) -> String {
-        format!("tcp://127.0.0.1:{}", self.port)
-    }
-
-    fn accepted(&self) -> usize {
-        self.accepted.load(Ordering::SeqCst)
-    }
-}
-
 #[tokio::test]
 async fn kerberos_kdc_down_fails_closed() -> anyhow::Result<()> {
     install_crypto_provider();
@@ -1341,170 +376,6 @@ async fn ntlm_injection_completes_credssp_both_legs() -> anyhow::Result<()> {
 
     let _ = gateway.process.start_kill();
     Ok(())
-}
-
-struct FakeClosedTarget {
-    port: u16,
-    accepted: Arc<AtomicUsize>,
-}
-
-impl FakeClosedTarget {
-    async fn start() -> anyhow::Result<Self> {
-        let listener = TcpListener::bind("127.0.0.1:0").await.context("bind closed target")?;
-        let port = listener.local_addr()?.port();
-        let accepted = Arc::new(AtomicUsize::new(0));
-        let accepted_task = Arc::clone(&accepted);
-        tokio::spawn(async move {
-            loop {
-                let Ok((_stream, _)) = listener.accept().await else {
-                    break;
-                };
-                accepted_task.fetch_add(1, Ordering::SeqCst);
-            }
-        });
-        Ok(Self { port, accepted })
-    }
-
-    fn accepted(&self) -> usize {
-        self.accepted.load(Ordering::SeqCst)
-    }
-}
-
-const IRONRDP_AGENT_VERSION: &str = "0.1.0";
-const RDCLEANPATH_INJECT_LOG: &str = "Switching to RdpProxy for credential injection (WebSocket)";
-const RDCLEANPATH_FORWARD_LOG: &str = "RDP-TLS forwarding (RDCleanPath)";
-
-fn ironrdp_agent_bin() -> Option<PathBuf> {
-    if let Ok(path) = std::env::var("IRONRDP_AGENT") {
-        return Some(PathBuf::from(path));
-    }
-    let name = if cfg!(windows) {
-        "ironrdp-agent.exe"
-    } else {
-        "ironrdp-agent"
-    };
-    if let Ok(home) = std::env::var("CARGO_HOME") {
-        let path = PathBuf::from(home).join("bin").join(name);
-        if path.is_file() {
-            return Some(path);
-        }
-    }
-    let cargo_home = std::env::var_os("USERPROFILE")
-        .or_else(|| std::env::var_os("HOME"))
-        .map(PathBuf::from)
-        .map(|home| home.join(".cargo").join("bin").join(name));
-    if let Some(path) = cargo_home
-        && path.is_file()
-    {
-        return Some(path);
-    }
-    if let Ok(path) = std::env::var("PATH") {
-        for dir in std::env::split_paths(&path) {
-            let candidate = dir.join(name);
-            if candidate.is_file() {
-                return Some(candidate);
-            }
-        }
-    }
-    None
-}
-
-fn require_ironrdp_agent() -> anyhow::Result<Option<PathBuf>> {
-    let Some(bin) = ironrdp_agent_bin() else {
-        eprintln!(
-            "skipping RDCleanPath ironrdp-agent test: cargo install ironrdp-agent --version {IRONRDP_AGENT_VERSION}"
-        );
-        return Ok(None);
-    };
-    let output = std::process::Command::new(&bin)
-        .arg("--version")
-        .output()
-        .with_context(|| format!("run {} --version", bin.display()))?;
-    let version = String::from_utf8_lossy(&output.stdout);
-    anyhow::ensure!(
-        version.contains(IRONRDP_AGENT_VERSION),
-        "expected ironrdp-agent {IRONRDP_AGENT_VERSION}, got {version:?} from {}",
-        bin.display()
-    );
-    Ok(Some(bin))
-}
-
-fn ironrdp_agent_endpoint() -> String {
-    let name = format!("ironrdp-e2e-{}", next_id().replace('-', ""));
-    if cfg!(windows) {
-        format!(r"\\.\pipe\{name}")
-    } else {
-        std::env::temp_dir().join(format!("{name}.sock")).display().to_string()
-    }
-}
-
-async fn start_ironrdp_daemon(bin: &Path, endpoint: &str) -> anyhow::Result<tokio::process::Child> {
-    let child = tokio::process::Command::new(bin)
-        .args(["--endpoint", endpoint, "daemon-start"])
-        .kill_on_drop(true)
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()
-        .context("start ironrdp-agent daemon")?;
-    let deadline = Instant::now() + Duration::from_secs(10);
-    loop {
-        let status = tokio::process::Command::new(bin)
-            .args(["--endpoint", endpoint, "status"])
-            .output()
-            .await
-            .context("ironrdp-agent status")?;
-        if status.status.success() {
-            return Ok(child);
-        }
-        if Instant::now() >= deadline {
-            anyhow::bail!(
-                "ironrdp-agent daemon not ready at {endpoint}: {}",
-                String::from_utf8_lossy(&status.stderr)
-            );
-        }
-        tokio::time::sleep(Duration::from_millis(50)).await;
-    }
-}
-
-async fn connect_ironrdp_rdcleanpath(
-    bin: &Path,
-    endpoint: &str,
-    server: &str,
-    token: &str,
-    http_port: u16,
-) -> anyhow::Result<tokio::process::Child> {
-    let url = format!("ws://127.0.0.1:{http_port}/jet/rdp");
-    tokio::process::Command::new(bin)
-        .args([
-            "--endpoint",
-            endpoint,
-            "connect",
-            "--server",
-            server,
-            "--username",
-            PROXY_USER,
-            "--password",
-            PROXY_PASSWORD,
-            "--prop",
-            &format!("ironrdp_rdcleanpathurl:s:{url}"),
-            "--prop",
-            &format!("ironrdp_rdcleanpathtoken:s:{token}"),
-        ])
-        .kill_on_drop(true)
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped())
-        .spawn()
-        .context("start ironrdp-agent connect")
-}
-
-async fn agent_query_logs(bin: &Path, endpoint: &str) -> String {
-    tokio::process::Command::new(bin)
-        .args(["--endpoint", endpoint, "query-logs"])
-        .output()
-        .await
-        .ok()
-        .map(|output| String::from_utf8_lossy(&output.stdout).into_owned())
-        .unwrap_or_default()
 }
 
 #[tokio::test]
